GameOver modal adaptive and other sounds were added.

+++ src/pages/Game.tsx
@@ -5,23 +5,86 @@ import { useStore } from '../store';
 import { observer } from 'mobx-react';
 import { GameOverModal } from '../components/GameOverModal';
 import useSound from 'use-sound';
-import nextRound from '../assets/sounds/next.mp3';
-import { Sandwich } from '../components/Sandwich';
 import classNames from 'classnames';
+import { Sandwich } from '../components/Sandwich';
+import nextRound from '../assets/sounds/next.mp3';
+import easyRound from '../assets/sounds/easy.mp3';
+import mediumRound from '../assets/sounds/medium.mp3';
+import hardRound from '../assets/sounds/hard.mp3';
+import hardRoundMillion from '../assets/sounds/hard_million.mp3';
+
+
 
 export const Game = observer(() => {
   const Store = useStore();
+
   const [modalVisible, setModalVisible] = useState(false);
   const [isWinningsOpen, setWinningsOpen] = useState(false);
 
-  const [playNextRound, { stop: stopPlayNextRound }] = useSound(
+  const [playNextRound, { stop: stopNextRound, isPlaying: isNextRoundPlaying }] = useSound(
       nextRound,
       { volume: 0.1 }
   );
 
+  const [playEasyRound, { stop: stopEasyRound }] = useSound(
+      easyRound,
+      { volume: 0.1 }
+  );
+
+  const [playMediumRound, { stop: stopMediumRound }] = useSound(
+      mediumRound,
+      { volume: 0.1 }
+  );
+
+  const [playHardRound, { stop: stopHardRound }] = useSound(
+      hardRound,
+      { volume: 0.1 }
+  );
+
+  const [playHardRoundMillion, { stop: stopHardRoundMillion }] = useSound(
+      hardRoundMillion,
+      { volume: 0.1 }
+  );
+
+  const stopPlayingRoundSounds = useCallback(
+      () => {
+        stopEasyRound();
+        stopMediumRound();
+        stopHardRound();
+        stopHardRoundMillion();
+      },
+      [stopEasyRound, stopMediumRound, stopHardRound, stopHardRoundMillion]
+  );
+
+  useEffect(() => {
+    if (Store.correctAnswer !== null || Store.isCheckingAnswer || Store.wrongAnswer !== null) {
+      stopPlayingRoundSounds();
+    }
+  }, [Store.correctAnswer, Store.isCheckingAnswer, Store.wrongAnswer, stopPlayingRoundSounds]);
+
+  useEffect(() => {
+    if (!isNextRoundPlaying) {
+      if (Store.currentQuestion?.id! < 4) {
+        playEasyRound();
+      }
+
+      if (Store.currentQuestion?.id! < 8 && Store.currentQuestion?.id! > 3) {
+        playMediumRound();
+      }
+
+      if (Store.currentQuestion?.id! < 11 && Store.currentQuestion?.id! > 7) {
+        playHardRound();
+      }
+
+      if (Store.currentQuestion?.id === 11) {
+        playHardRoundMillion();
+      }
+    }
+  }, [isNextRoundPlaying,Store.currentQuestion, playEasyRound, playMediumRound, playHardRound, playHardRoundMillion]);
+
   useEffect(() => {
     Store.getData();
-    Store.setCurrentQuestion(Store.questions[0]);
+    Store.setCurrentQuestion(Store.questions[7]);
   }, [Store]);
 
   useEffect(() => {
@@ -32,13 +95,22 @@ export const Game = observer(() => {
     }
   }, [Store.wrongAnswer]);
 
+
   useEffect(() => {
     if (Store.correctAnswer !== null && Store.wrongAnswer === null) {
-      setTimeout(() => {
-        playNextRound();
-        Store.setCurrentQuestion(Store.questions[Store.currentQuestion!.id + 1]);
-        Store.deleteAnswersHighlights();
-      }, 4000)
+      if (Store.currentQuestion?.id === 11) {
+        setTimeout(() => {
+          setModalVisible(true);
+          Store.setCorrectAnswer(null);
+        }, 5000)
+
+      } else {
+        setTimeout(() => {
+          playNextRound();
+          Store.setCurrentQuestion(Store.questions[Store.currentQuestion!.id + 1]);
+          Store.deleteAnswersHighlights();
+        }, 4000)
+      }
     }
   }, [Store,
     playNextRound,
@@ -60,7 +132,7 @@ export const Game = observer(() => {
       <section className="questions">
         <Sandwich isWinningsOpen={isWinningsOpen} toggleSandwich={toggleSandwich} />
         <h2 className="questions__heading">{Store.currentQuestion?.body}</h2>
-        <Answers stopPlayNextRound={stopPlayNextRound} />
+        <Answers stopNextRound={stopNextRound} stopPlayingRoundSounds={stopPlayingRoundSounds} />
       </section>
       <section className={winningsStyle}>
         <Winnings />
